+++ 32blit-stm32/Src/32blit.cpp
@@ -3,8 +3,8 @@
 #include <bitset>
 
 #include "32blit.h"
-#include "32blit_battery.hpp"
-#include "32blit_i2c.hpp"
+#include "32blit/battery.hpp"
+#include "32blit/i2c.hpp"
 #include "main.h"
 
 #include "adc.hpp"
