+++ 32blit-stm32/Src/32blit/battery.cpp
@@ -9,7 +9,7 @@
 //
 
 #include "32blit.h"
-#include "32blit_battery.hpp"
+#include "32blit/battery.hpp"
 
 using namespace blit;
