+++ 32blit-stm32/Src/SystemMenu/system_menu_controller.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "32blit.h"
-#include "32blit_battery.hpp"
+#include "32blit/battery.hpp"
 #include "32blit.hpp"
 #include <algorithm>
